Fix Windows warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@40895 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f5a6416949ea2be8c6bfa20ed2417707ed9203af

+++ Lbcom/Calo/CaloAssociators/src/CaloClusterMCTruth.cpp
@@ -1,4 +1,4 @@
-// $Id: CaloClusterMCTruth.cpp,v 1.11 2009-09-02 13:31:31 cattanem Exp $
+// $Id: CaloClusterMCTruth.cpp,v 1.12 2010-06-10 07:44:30 cattanem Exp $
 // ============================================================================
 // Include files 
 // ============================================================================
@@ -215,7 +215,7 @@ StatusCode CaloClusterMCTruth::execute    ()
         if ( cluster->e() > 0 && energy / cluster->e() < m_wFilter ) { continue; }
         
         // fill the relation table:         ATTENTION "i_push" is used!
-        table->i_push ( cluster , particle , energy ) ; // NB: "i_push"
+        table->i_push ( cluster , particle , (float)energy ) ; // NB: "i_push"
       }       
     }  // end of loop over clusters
   }  // end of loop over containers of clusters
